ayala-backend/requirements.txt
playwright==1.40.0
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
psycopg[binary]==3.2.9
pydantic==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
alembic==1.13.1
geoalchemy2==0.14.3
shapely>=2.0.2
python-dotenv==1.0.1
asyncpg==0.29.0
openai>=1.50.0
ruff==0.2.2
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.1
tenacity>=8.2.3
pydantic-settings==2.1.0
email-validator==2.1.0.post1
redis==5.0.1
aioredis==2.0.1

# Google OAuth
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1